Fix flakiness in `NMEAMessageHandlerTest`
This ensures that the test methods can never run out of order, since if they do it makes both of the first two fail if they run after either of the last two.  The renaming is needed since otherwise the methods would *always* run in the failing order, rather than never.

This was found and diagnosed using [iDFlakies](https://github.com/idflakies/iDFlakies).